LayoutTests/media/video-start.html

<video start=1s controls></video>
<script src=video-test.js></script>
<script>
    testExpected("video.start", 1.0);

    waitForEvent("canplaythrough", function () { testExpected("video.currentTime", 1.0); endTest(); });

    video.src = "content/test.mp4";
</script>
