LayoutTests/media/video-start-expected.txt

EXPECTED (video.start == '1') OK
EVENT(canplaythrough)
EXPECTED (video.currentTime == '1') OK
END OF TEST
